[build] fix `tt-xla` install
With latest `tt-xla` we have the `pjrt_plugin_tt` python package (which
encapsulates all shared code for the PJRT plugin). The python package
also ships with `pjrt_plugin_tt.so` shared library - note the same
name as the python package.

`tt-xla` gets installed in `install/` directory in `tt-torch`, so there
is a file `install/pjrt_plugin_tt.so`. Now, if you try to import
`pjrt_plugin_tt` in python, the python loader will first find the
`pjrt_plugin_tt.so` in `install/` and try to load it; however, this
shared object is not a python module - which causes a failuer.

Try to address this by not installing `tt-xla`, but just `tt-mlir` from
the `tt-xla` source tree. (The same thing is already being done for
`tt-metal` - i am not sure what was the reason for that)